requirements.txt
# general modules
django >3.2,<3.3
python-decouple
pillow
djangorestframework

# third party modules
markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt

# deployment modules

# email third party modules
django-mail-templated

# refactoring and tests
black
flake8